+++ webpack/static/scripts/config/pages.js
@@ -2,7 +2,8 @@
 
 var Pages = {
 	index: '.js-router--home',
-	example: '.js-router--example'
+	example: '.js-router--example',
+	'test/one': '.js-router--test-one'
 };
 
 module.exports = Pages;